+++ lib/cfg/instruction-sets/_all.ts
@@ -22,4 +22,4 @@
 // ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
 
-export {LlvmIrInstructionSetInfo} from './llvmir.js';
+export {LlvmIrInstructionSetInfo} from './llvm-ir.js';
